rename leet 75 study plan folder
