+++ study_plans/leet_75/sort_list/sort_linked_list.py
